+++ src/envoydevice.js
@@ -31,25 +31,40 @@ class EnvoyDevice extends EventEmitter {
         this.envoySerialNumber = config.envoySerialNumber || '';
         this.enlightenUser = config.enlightenUser;
         this.enlightenPassword = config.enlightenPasswd;
+
         this.powerProductionSummary = config.powerProductionSummary || 0;
-        this.powerProductionStateSensor = config.powerProductionOnOff || false;
-        this.powerProductionPeakSensor = config.powerProductionMax || false;
-        this.powerProductionPeakAutoReset = config.powerProductionMaxAutoReset || 0;
-        this.powerProductionPeakDetectedValue = config.powerProductionMaxDetected || 0;
-        this.powerConsumptionTotalPowerPeakSensor = config.powerConsumptionTotalMax || false;
-        this.powerConsumptionTotalPowerPeakAutoReset = config.powerConsumptionTotalMaxAutoReset || 0;
-        this.powerConsumptionTotalPowerPeakDetected = config.powerConsumptionTotalMaxDetected || 0;
+        this.powerProductionStateSensor = config.powerProductionStateSensor || false;
+        this.powerProductionPeakSensor = config.powerProductionPeakSensor || false;
+        this.powerProductionPeakSensorDetected = config.powerProductionPeakSensorDetected || 0;
+        this.powerProductionPeakAutoReset = config.powerProductionPeakAutoReset || 0;
         this.energyProductionStateSensor = config.energyProductionStateSensor || false;
         this.energyProductionLevelSensor = config.energyProductionLevelSensor || false;
-        this.energyProductionLevelDetectedValue = config.energyProductionLevelDetected || 0;
+        this.energyProductionLevelSensorDetected = config.energyProductionLevelSensorDetected || 0;
         this.energyProductionLifetimeOffset = config.energyProductionLifetimeOffset || 0;
+
+        this.powerConsumptionTotalStateSensor = config.powerConsumptionTotalStateSensor || false;
+        this.powerConsumptionTotalPeakSensor = config.powerConsumptionTotalPeakSensor || false;
+        this.powerConsumptionTotalPeakSensorDetected = config.powerConsumptionTotalPeakSensorDetected || 0;
+        this.powerConsumptionTotalPeakAutoReset = config.powerConsumptionTotalPeakAutoReset || 0;
+        this.energyConsumptionTotaStateSensor = config.energyConsumptionTotaStateSensor || false;
+        this.energyConsumptionTotaLevelSensor = config.energyConsumptionTotaLevelSensor || false;
+        this.energyConsumptionTotaLevelSensorDetected = config.energyConsumptionTotaLevelSensorDetected || 0;
         this.energyConsumptionTotalLifetimeOffset = config.energyConsumptionTotalLifetimeOffset || 0;
-        this.powerConsumptionNetPowerPeakAutoReset = config.powerConsumptionNetMaxAutoReset || 0;
-        this.powerConsumptionNetPowerPeakSensor = config.powerConsumptionNetMax || false;
-        this.powerConsumptionNetPowerPeakDetected = config.powerConsumptionNetMaxDetected || 0;
+
+        this.powerConsumptionNetStateSensor = config.powerConsumptionNetStateSensor || false;
+        this.powerConsumptionNetPeakSensor = config.powerConsumptionNetPeakSensor || false;
+        this.powerConsumptionNetPeakSensorDetected = config.powerConsumptionNetPeakSensorDetected || 0;
+        this.powerConsumptionNetPeakAutoReset = config.powerConsumptionNetPeakAutoReset || 0;
+        this.energyConsumptionNetStateSensor = config.energyConsumptionNetStateSensor || false;
+        this.energyConsumptionNetLevelSensor = config.energyConsumptionNetLevelSensor || false;
+        this.energyConsumptionNetLevelSensorDetected = config.energyConsumptionNetLevelSensorDetected || 0;
         this.energyConsumptionNetLifetimeOffset = config.energyConsumptionNetLifetimeOffset || 0;
-        this.supportEnsembleStatus = this.envoyFirmware7xx ? config.supportEnsembleStatus : false || false;
-        this.supportLiveData = this.envoyFirmware7xx ? config.supportLiveData : false || false;
+
+        this.supportEnsembleStatus = this.envoyFirmware7xx ? config.supportEnsembleStatus : false;
+        this.enpowerGridStateSensor = this.supportEnsembleStatus ? config.enpowerGridStateSensor : false;
+        this.enchargeGridStateSensor = this.supportEnsembleStatus ? config.enchargeGridStateSensor : false;
+        this.solarGridStateSensor = this.supportEnsembleStatus ? config.solarGridStateSensor : false;
+        this.supportLiveData = this.envoyFirmware7xx ? config.supportLiveData : false;
         this.liveDataRefreshTime = config.liveDataRefreshTime || 1000;
         this.metersDataRefreshTime = config.metersDataRefreshTime || 2500;
         this.productionDataRefreshTime = config.productionDataRefreshTime || 5000;
@@ -64,7 +79,7 @@ class EnvoyDevice extends EventEmitter {
         this.mqttEnabled = config.enableMqtt || false;
         this.mqttHost = config.mqttHost;
         this.mqttPort = config.mqttPort || 1883;
-        this.mqttClientId = config.mqttClientId || `mqtt_${Math.random().toString(16).slice(3)}`;
+        this.mqttClientId = config.mqttClientId || `envoy_${Math.random().toString(16).slice(3)}`;
         this.mqttPrefix = config.mqttPrefix;
         this.mqttAuth = config.mqttAuth || false;
         this.mqttUser = config.mqttUser;
@@ -125,12 +140,13 @@ class EnvoyDevice extends EventEmitter {
         this.envoyCommEnchagLevelSubg = 0;
         this.envoyEnpowerConnected = false;
         this.envoyEnpowerGridStatus = 'Unknown';
+        this.envoyEnpowerGridState = false;
 
         //microinverters
         this.microinvertersSupported = false;
         this.microinvertersInstalled = false;
         this.microinvertersCount = 0;
-        this.microinvertersType = 'Unknown'
+        this.microinvertersType = 'Unknown';
 
         //qrelay
         this.qRelaysSupported = false;
@@ -142,26 +158,26 @@ class EnvoyDevice extends EventEmitter {
         this.acBatteriesSupported = false;
         this.acBatteriesInstalled = false;
         this.acBatteriesCount = 0;
-        this.acBatteriesType = 'Unknown'
-
+        this.acBatteriesType = 'Unknown';
 
         //ensemble esub
         this.esubsSupported = false;
         this.esubsInstalled = false;
         this.esubsCount = 0;
-        this.esubsType = 'Unknown'
+        this.esubsType = 'Unknown';
 
         //encharges
         this.enchargesSupported = false;
         this.enchargesInstalled = false;
         this.enchargesCount = 0;
-        this.enchargesType = 'Unknown'
+        this.enchargesType = 'Unknown';
 
         //enpower
         this.enpowersSupported = false;
         this.enpowersInstalled = false;
         this.enpowersCount = 0;
-        this.enpowersType = ''
+        this.enpowersType = '';
+        this.enchargesGridState = false;
 
         //ct meters
         this.metersSupported = false;
@@ -203,8 +219,14 @@ class EnvoyDevice extends EventEmitter {
         this.productionReadingTime = '';
 
         //consumption CT
+        this.consumptionTotalPowerState = false;
         this.consumptionTotalPowerPeakDetected = false;
+        this.consumptionTotalEnergyState = false;
+        this.consumptionTotalEnergyLevelDetected = false;
+        this.consumptionNetEnergyState = false;
+        this.consumptionNetPowerState = false;
         this.consumptionNetPowerPeakDetected = false;
+        this.consumptionNetEnergyLevelDetected = false;
 
         //ac batteries summary
         this.acBatteriesSummaryType = 'Unknown';
@@ -242,6 +264,9 @@ class EnvoyDevice extends EventEmitter {
         this.ensembleGridProfileVersion = 'Unknown';
         this.ensembleItemCount = 0;
         this.ensembleFakeInventoryMode = false;
+        this.enchargeGridState = false;
+        this.solarGridState = false;
+
         this.url = this.envoyFirmware7xx ? `https://${this.host}` : `http://${this.host}`;
 
         //current day of week/month
@@ -835,6 +860,7 @@ class EnvoyDevice extends EventEmitter {
                 const enpower = enpowersSupported ? envoy.enpower : {};
                 const enpowerConnected = enpowersSupported ? enpower.connected === true : false;
                 const enpowerGridStatus = enpowersSupported ? CONSTANS.ApiCodes[enpower.grid_status] || 'undefined' : '';
+                const enpowerGridState = enpowerGridStatus !== 'Enpower state no grid';
 
                 //convert status
                 const status = (Array.isArray(alerts) && alerts.length > 0) ? (alerts.map(a => CONSTANS.ApiCodes[a] || a).join(', ')).substring(0, 64) : 'No alerts';
@@ -866,6 +892,12 @@ class EnvoyDevice extends EventEmitter {
                             .updateCharacteristic(Characteristic.enphaseEnvoyEnpowerConnected, enpowerConnected)
                             .updateCharacteristic(Characteristic.enphaseEnvoyEnpowerGridStatus, enpowerGridStatus)
                     }
+                }
+
+                //enpower grid state sensors
+                if (this.enpowerGridStateSensorService) {
+                    this.enpowerGridStateSensorService
+                        .updateCharacteristic(Characteristic.ContactSensorState, enpowerGridState)
                 }
 
                 this.microinvertersSupported = microinvertersSupported;
@@ -906,6 +938,7 @@ class EnvoyDevice extends EventEmitter {
 
                 this.envoyEnpowerConnected = enpowerConnected;
                 this.envoyEnpowerGridStatus = enpowerGridStatus;
+                this.envoyEnpowerGridState = enpowerGridState;
 
                 //restFul
                 const restFul = this.restFulConnected ? this.restFul.update('home', homeData.data) : false;
@@ -1920,6 +1953,23 @@ class EnvoyDevice extends EventEmitter {
                 const der2State = relay.der2_state;
                 const enchGridMode = CONSTANS.ApiCodes[relay.Enchg_grid_mode] || 'unknown';
                 const solarGridMode = CONSTANS.ApiCodes[relay.Solar_grid_mode] || 'unknown';
+                const enchargeGridState = enchGridMode !== 'Enpower state no grid';
+                const solarGridState = solarGridMode !== 'Enpower state no grid';
+
+                //encharge grid state sensors
+                if (this.enchargeGridStateSensorService) {
+                    this.enchargeGridStateSensorService
+                        .updateCharacteristic(Characteristic.ContactSensorState, enchargeGridState)
+                }
+
+                //solar grid state sensors
+                if (this.solarGridStateSensorService) {
+                    this.solarGridStateSensorService
+                        .updateCharacteristic(Characteristic.ContactSensorState, solarGridState)
+                }
+
+                this.enchargeGridState = enchargeGridState;
+                this.solarGridState = solarGridState;
 
                 //profile
                 const profile = ensembleStatus.profile;
@@ -2178,9 +2228,9 @@ class EnvoyDevice extends EventEmitter {
             const date = new Date();
             const currentDayOfWeek = date.getDay();
             const currentDayOfMonth = date.getDate();
-            const resetProductionPowerPeak = [false, currentDayOfWeek !== this.currentDayOfWeek, currentDayOfWeek === 6 ? currentDayOfWeek < this.currentDayOfWeek : false, currentDayOfMonth < this.currentDayOfMonth][this.powerProductionPowerPeakAutoReset];
-            const resetConsumptionTotalPowerPeak = [false, currentDayOfWeek !== this.currentDayOfWeek, currentDayOfWeek === 6 ? currentDayOfWeek < this.currentDayOfWeek : false, currentDayOfMonth < this.currentDayOfMonth][this.powerConsumptionTotalPowerPeakAutoReset];
-            const resetConsumptionNetPowerPeak = [false, currentDayOfWeek !== this.currentDayOfWeek, currentDayOfWeek === 6 ? currentDayOfWeek < this.currentDayOfWeek : false, currentDayOfMonth < this.currentDayOfMonth][this.powerConsumptionNetPowerPeakAutoReset];
+            const resetProductionPowerPeak = [false, currentDayOfWeek !== this.currentDayOfWeek, currentDayOfWeek === 6 ? currentDayOfWeek < this.currentDayOfWeek : false, currentDayOfMonth < this.currentDayOfMonth][this.powerProductionPeakAutoReset];
+            const resetConsumptionTotalPowerPeak = [false, currentDayOfWeek !== this.currentDayOfWeek, currentDayOfWeek === 6 ? currentDayOfWeek < this.currentDayOfWeek : false, currentDayOfMonth < this.currentDayOfMonth][this.powerConsumptionTotalPeakAutoReset];
+            const resetConsumptionNetPowerPeak = [false, currentDayOfWeek !== this.currentDayOfWeek, currentDayOfWeek === 6 ? currentDayOfWeek < this.currentDayOfWeek : false, currentDayOfMonth < this.currentDayOfMonth][this.powerConsumptionNetPeakAutoReset];
 
             //get enabled devices
             const metersProductionEnabled = this.metersProductionEnabled;
@@ -2210,11 +2260,11 @@ class EnvoyDevice extends EventEmitter {
                 const productionReadingTime = metersProductionEnabled ? new Date(production.readingTime * 1000).toLocaleString() : productionMicroReadingTime;
                 const productionPower = metersProductionEnabled ? parseFloat(production.wNow) / 1000 : this.productionMicroSummaryWattsNow;
 
-                //production power state
+                //power state
                 const productionPowerState = productionPower > 0; // true if power > 0
                 const debug1 = this.enableDebugMode ? this.emit('debug', `Production power state: ${productionPowerState}`) : false;
 
-                //production power level
+                //power level
                 const powerProductionSummary = this.powerProductionSummary / 1000; //kW
                 const productionPowerLevel = productionPowerState ? (Math.min(Math.max((100 / powerProductionSummary) * productionPower, 1), 100)).toFixed(1) : 0; //0-100%
                 const debug2 = this.enableDebugMode ? this.emit('debug', `Production power level: ${productionPowerLevel} %`) : false;
@@ -2229,8 +2279,8 @@ class EnvoyDevice extends EventEmitter {
                 const write = (productionPower > productionPowerPeak) || resetProductionPowerPeak ? await fsPromises.writeFile(this.productionPowerPeakFile, powerProductionToWrite) : false;
                 const showLog = write && this.enableDebugMode ? this.emit('debug', `Saved production power peak: ${powerProductionToWrite} kW`) : false;
 
-                //power peak state detected
-                const productionPowerPeakDetected = productionPower >= (this.powerProductionPeakDetectedValue / 1000);
+                //power peak state
+                const productionPowerPeakDetected = productionPower >= (this.powerProductionPeakSensorDetected / 1000);
                 const debug4 = this.enableDebugMode ? this.emit('debug', `Production power peak detected: ${productionPowerPeakDetected}`) : false;
 
                 //energy
@@ -2246,12 +2296,12 @@ class EnvoyDevice extends EventEmitter {
                 //energy lifetime fix for negative value
                 const productionEnergyLifeTimeFix = Math.min(productionEnergyLifeTime, 0);
 
-                //production energy state
+                //energy state
                 const productionEnergyState = productionEnergyToday > 0; // true if energy > 0
                 const debug5 = this.enableDebugMode ? this.emit('debug', `Production energy level state: ${productionEnergyState}`) : false;
 
-                //production energy level detected
-                const productionEnergyLevelDetected = productionEnergyToday >= (this.energyProductionLevelDetectedValue / 1000);
+                //energy level
+                const productionEnergyLevelDetected = productionEnergyToday >= (this.energyProductionLevelSensorDetected / 1000);
                 const debug6 = this.enableDebugMode ? this.emit('debug', `Production energy level detected: ${productionEnergyLevelDetected}`) : false;
 
                 //param
@@ -2358,6 +2408,10 @@ class EnvoyDevice extends EventEmitter {
                         const consumptionReadingTime = new Date(consumption.readingTime * 1000).toLocaleString();
                         const consumptionPower = parseFloat(consumption.wNow) / 1000;
 
+                        //power state
+                        const consumptionPowerState = consumptionPower > 0; // true if power > 0
+                        const debug1 = this.enableDebugMode ? this.emit('debug', `Consumption ${['Total', 'Net'][i]} power state: ${consumptionPowerState}`) : false;
+
                         //read saved power peak
                         const consumptionsName = ['consumption total', 'consumption net'][i];
                         const consumptionsFile = [this.consumptionTotalPowerPeakFile, this.consumptionNetPowerPeakFile][i];
@@ -2371,8 +2425,8 @@ class EnvoyDevice extends EventEmitter {
                         const write = (consumptionPower > consumptionPowerPeak) || autoReset ? await fsPromises.writeFile(consumptionsFile, consumptionPowerToWrite) : false;
                         const showLog = write && this.enableDebugMode ? this.emit('debug', `Saved ${consumptionsName} power peak: ${consumptionPowerToWrite} kW`) : false;
 
-                        //power peak state detected
-                        const consumptionsPowerPeakDetected = [this.powerConsumptionTotalPowerPeakDetected / 1000, this.powerConsumptionNetPowerPeakDetected / 1000][i];
+                        //power peak state
+                        const consumptionsPowerPeakDetected = [this.powerConsumptionTotalPeakSensorDetected / 1000, this.powerConsumptionNetPeakSensorDetected / 1000][i];
                         const consumptionPowerPeakDetected = consumptionPower >= consumptionsPowerPeakDetected || false;
 
                         //energy
@@ -2385,6 +2439,15 @@ class EnvoyDevice extends EventEmitter {
                         const consumptionEnergyVahToday = parseFloat(consumption.vahToday) / 1000;
                         const consumptionEnergyVarhLeadToday = parseFloat(consumption.varhLeadToday) / 1000;
                         const consumptionEnergyVarhLagToday = parseFloat(consumption.varhLagToday) / 1000;
+
+                        //energy state
+                        const consumptionEnergyState = consumptionEnergyToday > 0; // true if energy > 0
+                        const debug5 = this.enableDebugMode ? this.emit('debug', `Consumption ${['Total', 'Net'][i]} energy level state: ${consumptionEnergyState}`) : false;
+
+                        //energy level 
+                        const consumptionsEnergyLevelDetected = [this.energyConsumptionTotalLevelSensorDetected / 1000, this.energyConsumptionNetLevelSensorDetected / 1000][i];
+                        const consumptionEnergyLevelDetected = consumptionEnergyToday >= consumptionsEnergyLevelDetected || false;
+                        const debug6 = this.enableDebugMode ? this.emit('debug', `Consumption ${['Total', 'Net'][i]} energy level detected: ${consumptionEnergyLevelDetected}`) : false;
 
                         //net param
                         const consumptionRmsCurrent = parseFloat(consumption.rmsCurrent);
@@ -2427,20 +2490,58 @@ class EnvoyDevice extends EventEmitter {
                         this.consumptionsApparentPower.push(consumptionApparentPower);
                         this.consumptionsPwrFactor.push(consumptionPwrFactor);
 
-                        const consumptionTotalPowerPeakDetected = i === 0 ? this.consumptionsPowerPeakDetected[i] : false;
-                        const consumptionNetPowerPeakDetected = i === 1 ? this.consumptionsPowerPeakDetected[i] : false;
+                        if (i === 0) {
+                            if (this.consumptionTotalPowerStateSensorService) {
+                                this.consumptionTotalPowerStateSensorService
+                                    .updateCharacteristic(Characteristic.ContactSensorState, consumptionPowerState)
+                            }
 
-                        if (this.consumptionTotalPowerPeakSensorService) {
-                            this.consumptionTotalPowerPeakSensorService
-                                .updateCharacteristic(Characteristic.ContactSensorState, consumptionTotalPowerPeakDetected)
+                            if (this.consumptionTotalPowerPeakSensorService) {
+                                this.consumptionTotalPowerPeakSensorService
+                                    .updateCharacteristic(Characteristic.ContactSensorState, consumptionPowerPeakDetected)
+                            }
+
+                            if (this.consumptionTotalEnergyyStateSensorService) {
+                                this.consumptionTotalEnergyyStateSensorService
+                                    .updateCharacteristic(Characteristic.ContactSensorState, consumptionEnergyState)
+                            }
+
+                            if (this.consumptionTotalEnergyLevelSensorService) {
+                                this.consumptionTotalEnergyLevelSensorService
+                                    .updateCharacteristic(Characteristic.ContactSensorState, consumptionEnergyLevelDetected)
+                            }
+                            this.consumptionTotalPowerState = consumptionPowerState;
+                            this.consumptionTotalPowerPeakDetected = consumptionPowerPeakDetected;
+                            this.consumptionTotalEnergyState = consumptionEnergyState;
+                            this.consumptionTotalEnergyLevelDetected = consumptionEnergyLevelDetected;
                         }
 
-                        if (this.consumptionNetPowerPeakSensorService) {
-                            this.consumptionNetPowerPeakSensorService
-                                .updateCharacteristic(Characteristic.ContactSensorState, consumptionNetPowerPeakDetected)
+                        if (i === 1) {
+                            if (this.consumptionNetPowerStateSensorService) {
+                                this.consumptionNetPowerStateSensorService
+                                    .updateCharacteristic(Characteristic.ContactSensorState, consumptionPowerState)
+                            }
+
+                            if (this.consumptionNetPowerPeakSensorService) {
+                                this.consumptionNetPowerPeakSensorService
+                                    .updateCharacteristic(Characteristic.ContactSensorState, consumptionPowerPeakDetected)
+                            }
+
+                            if (this.consumptionNetEnergyyStateSensorService) {
+                                this.consumptionNetEnergyyStateSensorService
+                                    .updateCharacteristic(Characteristic.ContactSensorState, consumptionEnergyState)
+                            }
+
+                            if (this.consumptionNetEnergyLevelSensorService) {
+                                this.consumptionNetEnergyLevelSensorService
+                                    .updateCharacteristic(Characteristic.ContactSensorState, consumptionEnergyLevelDetected)
+                            }
+                            this.consumptionNetPowerState = consumptionPowerState;
+                            this.consumptionNetPowerPeakDetected = consumptionPowerPeakDetected;
+                            this.consumptionNetEnergyState = consumptionEnergyState;
+                            this.consumptionNetEnergyLevelDetected = consumptionEnergyLevelDetected;
                         }
-                        this.consumptionTotalPowerPeakDetected = consumptionTotalPowerPeakDetected;
-                        this.consumptionNetPowerPeakDetected = consumptionNetPowerPeakDetected;
+
                         this.metersConsumptionCount = metersConsumptionCount;
                     }
                 }
@@ -2982,6 +3083,20 @@ class EnvoyDevice extends EventEmitter {
                 this.envoysService.push(enphaseEnvoyService);
                 accessory.addService(this.envoysService[0]);
 
+                //grid state sensor service
+                if (this.enpowerGridStateSensor) {
+                    const debug = this.enableDebugMode ? this.emit('debug', `Prepare envoy enpower grid state sensor service`) : false;
+                    this.enpowerGridStateSensorService = new Service.ContactSensor(`${accessoryName} Enpower Grid State`, `Enpower Grid State`);
+                    this.enpowerGridStateSensorService.getCharacteristic(Characteristic.ConfiguredName, `${accessoryName} Enpower Grid State`);
+                    this.enpowerGridStateSensorService.getCharacteristic(Characteristic.ContactSensorState)
+                        .onGet(async () => {
+                            const state = this.envoyEnpowerGridState = false;;
+                            const info = this.disableLogInfo ? false : this.emit('message', `Enpower grid state sensor: ${state ? 'Active' : 'Not active'}`);
+                            return state;
+                        });
+                    accessory.addService(this.enpowerGridStateSensorService);
+                };
+
                 //qrelays
                 if (qRelaysInstalled) {
                     this.qRelaysService = [];
@@ -3266,7 +3381,7 @@ class EnvoyDevice extends EventEmitter {
                 this.productionsService.push(enphaseProductionService);
                 accessory.addService(this.productionsService[0]);
 
-                //prepare production state sensor service
+                //production state sensor service
                 if (this.powerProductionStateSensor) {
                     const debug = this.enableDebugMode ? this.emit('debug', `Prepare production power state sensor service`) : false;
                     this.productionPowerStateSensorService = new Service.ContactSensor(`${accessoryName} Production Power State`, `Production Power State`);
@@ -3280,7 +3395,7 @@ class EnvoyDevice extends EventEmitter {
                     accessory.addService(this.productionPowerStateSensorService);
                 };
 
-                //prepare production power peak sensor service
+                //production power peak sensor service
                 if (this.powerProductionPeakSensor) {
                     const debug = this.enableDebugMode ? this.emit('debug', `Prepare production power peak sensor service`) : false;
                     this.productionPowerPeakSensorService = new Service.ContactSensor(`${accessoryName} Production Power Peak`, `Production Power Peak`);
@@ -3294,7 +3409,7 @@ class EnvoyDevice extends EventEmitter {
                     accessory.addService(this.productionPowerPeakSensorService);
                 };
 
-                //prepare production energy state sensor service
+                //production energy state sensor service
                 if (this.energyProductionStateSensor) {
                     const debug = this.enableDebugMode ? this.emit('debug', `Prepare production energy state sensor service`) : false;
                     this.productionEnergyStateSensorService = new Service.ContactSensor(`${accessoryName} Production Energy State`, `Production Energy State`);
@@ -3308,7 +3423,7 @@ class EnvoyDevice extends EventEmitter {
                     accessory.addService(this.productionEnergyStateSensorService);
                 };
 
-                //prepare production energy level sensor service
+                //production energy level sensor service
                 if (this.energyProductionLevelSensor) {
                     const debug = this.enableDebugMode ? this.emit('debug', `Prepare production energy level sensor service`) : false;
                     this.productionEnergyLevelSensorService = new Service.ContactSensor(`${accessoryName} Production Energy Level`, `Production Energy Level`);
@@ -3420,32 +3535,122 @@ class EnvoyDevice extends EventEmitter {
                         this.consumptionsService.push(enphaseConsumptionService);
                         accessory.addService(this.consumptionsService[i]);
 
-                        //prepare consumption total power peak sensor service
-                        if (this.powerConsumptionTotalPowerPeakSensor && i === 0) {
-                            const debug = this.enableDebugMode ? this.emit('debug', `Prepare consumption total power peak service`) : false;
-                            this.consumptionTotalPowerPeakSensorService = new Service.ContactSensor(`${accessoryName} Power Peak Total`, `Power Peak Total`);
-                            this.consumptionTotalPowerPeakSensorService.getCharacteristic(Characteristic.ConfiguredName, `${accessoryName} Power Peak Total`);
-                            this.consumptionTotalPowerPeakSensorService.getCharacteristic(Characteristic.ContactSensorState)
-                                .onGet(async () => {
-                                    const state = this.consumptionTotalPowerPeakDetected || false;
-                                    const info = this.disableLogInfo ? false : this.emit('message', `${consumptionMeasurmentType}, power peak sensor: ${state ? 'Active' : 'Not active'}`);
-                                    return state;
-                                });
-                            accessory.addService(this.consumptionTotalPowerPeakSensorService)
+                        //total
+                        if (i === 0) {
+                            //consumption total state sensor service
+                            if (this.powerConsumptionTotalStateSensor) {
+                                const debug = this.enableDebugMode ? this.emit('debug', `Prepare consumption total power state sensor service`) : false;
+                                this.consumptionTotalPowerStateSensorService = new Service.ContactSensor(`${accessoryName} Consumption Total Power State`, `Consumption Total Power State`);
+                                this.consumptionTotalPowerStateSensorService.getCharacteristic(Characteristic.ConfiguredName, `${accessoryName} Consumption Total Power State`);
+                                this.consumptionTotalPowerStateSensorService.getCharacteristic(Characteristic.ContactSensorState)
+                                    .onGet(async () => {
+                                        const state = this.consumptionTotalPowerState;
+                                        const info = this.disableLogInfo ? false : this.emit('message', `Consumption Total power state sensor: ${state ? 'Active' : 'Not active'}`);
+                                        return state;
+                                    });
+                                accessory.addService(this.consumptionTotalPowerStateSensorService);
+                            };
+
+                            //consumption total power peak sensor service
+                            if (this.powerConsumptionTotalPeakSensor) {
+                                const debug = this.enableDebugMode ? this.emit('debug', `Prepare consumption total power peak service`) : false;
+                                this.consumptionTotalPowerPeakSensorService = new Service.ContactSensor(`${accessoryName} Power Peak Total`, `Power Peak Total`);
+                                this.consumptionTotalPowerPeakSensorService.getCharacteristic(Characteristic.ConfiguredName, `${accessoryName} Power Peak Total`);
+                                this.consumptionTotalPowerPeakSensorService.getCharacteristic(Characteristic.ContactSensorState)
+                                    .onGet(async () => {
+                                        const state = this.consumptionTotalPowerPeakDetected;
+                                        const info = this.disableLogInfo ? false : this.emit('message', `${consumptionMeasurmentType}, power peak sensor: ${state ? 'Active' : 'Not active'}`);
+                                        return state;
+                                    });
+                                accessory.addService(this.consumptionTotalPowerPeakSensorService)
+                            };
+
+                            //consumption total energy state sensor service
+                            if (this.energyConsumptionTotalStateSensor) {
+                                const debug = this.enableDebugMode ? this.emit('debug', `Prepare consumption total energy state sensor service`) : false;
+                                this.consumptionTotalEnergyStateSensorService = new Service.ContactSensor(`${accessoryName} Consumption Total Energy State`, `Consumption Total Energy State`);
+                                this.consumptionTotalEnergyStateSensorService.getCharacteristic(Characteristic.ConfiguredName, `${accessoryName} Consumption Total Energy State`);
+                                this.consumptionTotalEnergyStateSensorService.getCharacteristic(Characteristic.ContactSensorState)
+                                    .onGet(async () => {
+                                        const state = this.consumptionTotalEnergyState;
+                                        const info = this.disableLogInfo ? false : this.emit('message', `Consumption Total energy state sensor: ${state ? 'Active' : 'Not active'}`);
+                                        return state;
+                                    });
+                                accessory.addService(this.consumptionTotalEnergyStateSensorService);
+                            };
+
+                            //consumption total energy level sensor service
+                            if (this.energyConsumptionTotaLevelSensor) {
+                                const debug = this.enableDebugMode ? this.emit('debug', `Prepare consumption total energy level sensor service`) : false;
+                                this.consumptionTotalEnergyLevelSensorService = new Service.ContactSensor(`${accessoryName} Consumption Total Energy Level`, `Consumption Total Energy Level`);
+                                this.consumptionTotalEnergyLevelSensorService.getCharacteristic(Characteristic.ConfiguredName, `${accessoryName} Consumption Total Energy Level`);
+                                this.consumptionTotalEnergyLevelSensorService.getCharacteristic(Characteristic.ContactSensorState)
+                                    .onGet(async () => {
+                                        const state = this.consumptionTotalEnergyLevelDetected;
+                                        const info = this.disableLogInfo ? false : this.emit('message', `Consumption total energy level sensor: ${state ? 'Active' : 'Not active'}`);
+                                        return state;
+                                    });
+                                accessory.addService(this.consumptionTotalEnergyLevelSensorService);
+                            };
                         };
 
-                        //prepare consumption net power peak sensor service
-                        if (this.powerConsumptionNetPowerPeakSensor && i === 1) {
-                            const debug = this.enableDebugMode ? this.emit('debug', `Prepare consumption net power peak sensor service`) : false;
-                            this.consumptionNetPowerPeakSensorService = new Service.ContactSensor(`${accessoryName} Power Peak Net`, `Power Peak Net`);
-                            this.consumptionNetPowerPeakSensorService.getCharacteristic(Characteristic.ConfiguredName, `${accessoryName} Power Peak Net`);
-                            this.consumptionNetPowerPeakSensorService.getCharacteristic(Characteristic.ContactSensorState)
-                                .onGet(async () => {
-                                    const state = this.consumptionNetPowerPeakDetected || false;
-                                    const info = this.disableLogInfo ? false : this.emit('message', `${consumptionMeasurmentType}, power peak sensor: ${state ? 'Active' : 'Not active'}`);
-                                    return state;
-                                });
-                            accessory.addService(this.consumptionNetPowerPeakSensorService)
+                        //net
+                        if (i === 1) {
+                            //consumption total state sensor service
+                            if (this.powerConsumptionNetStateSensor) {
+                                const debug = this.enableDebugMode ? this.emit('debug', `Prepare consumption net power state sensor service`) : false;
+                                this.consumptionNetPowerStateSensorService = new Service.ContactSensor(`${accessoryName} Consumption Net Power State`, `Consumption Net Power State`);
+                                this.consumptionNetPowerStateSensorService.getCharacteristic(Characteristic.ConfiguredName, `${accessoryName} Consumption Net Power State`);
+                                this.consumptionNetPowerStateSensorService.getCharacteristic(Characteristic.ContactSensorState)
+                                    .onGet(async () => {
+                                        const state = this.consumptionNetPowerState;
+                                        const info = this.disableLogInfo ? false : this.emit('message', `Consumption Net power state sensor: ${state ? 'Active' : 'Not active'}`);
+                                        return state;
+                                    });
+                                accessory.addService(this.consumptionNetPowerStateSensorService);
+                            };
+
+                            //consumption net power peak sensor service
+                            if (this.powerConsumptionNetPeakSensor) {
+                                const debug = this.enableDebugMode ? this.emit('debug', `Prepare consumption net power peak sensor service`) : false;
+                                this.consumptionNetPowerPeakSensorService = new Service.ContactSensor(`${accessoryName} Power Peak Net`, `Power Peak Net`);
+                                this.consumptionNetPowerPeakSensorService.getCharacteristic(Characteristic.ConfiguredName, `${accessoryName} Power Peak Net`);
+                                this.consumptionNetPowerPeakSensorService.getCharacteristic(Characteristic.ContactSensorState)
+                                    .onGet(async () => {
+                                        const state = this.consumptionNetPowerPeakDetected;
+                                        const info = this.disableLogInfo ? false : this.emit('message', `${consumptionMeasurmentType}, power peak sensor: ${state ? 'Active' : 'Not active'}`);
+                                        return state;
+                                    });
+                                accessory.addService(this.consumptionNetPowerPeakSensorService)
+                            };
+
+                            //consumption net energy state sensor service
+                            if (this.energyConsumptionNetStateSensor) {
+                                const debug = this.enableDebugMode ? this.emit('debug', `Prepare consumption net energy state sensor service`) : false;
+                                this.consumptionNetEnergyStateSensorService = new Service.ContactSensor(`${accessoryName} Consumption Net Energy State`, `Consumption Net Energy State`);
+                                this.consumptionNetEnergyStateSensorService.getCharacteristic(Characteristic.ConfiguredName, `${accessoryName} Consumption Net Energy State`);
+                                this.consumptionNetEnergyStateSensorService.getCharacteristic(Characteristic.ContactSensorState)
+                                    .onGet(async () => {
+                                        const state = this.consumptionNetEnergyState;
+                                        const info = this.disableLogInfo ? false : this.emit('message', `Consumption Net energy state sensor: ${state ? 'Active' : 'Not active'}`);
+                                        return state;
+                                    });
+                                accessory.addService(this.consumptionNetEnergyStateSensorService);
+                            };
+
+                            //consumption net energy level sensor service
+                            if (this.energyConsumptionNetLevelSensor) {
+                                const debug = this.enableDebugMode ? this.emit('debug', `Prepare consumption net energy level sensor service`) : false;
+                                this.consumptionNetEnergyLevelSensorService = new Service.ContactSensor(`${accessoryName} Consumption Net Energy Level`, `Consumption Net Energy Level`);
+                                this.consumptionNetEnergyLevelSensorService.getCharacteristic(Characteristic.ConfiguredName, `${accessoryName} Consumption Net Energy Level`);
+                                this.consumptionNetEnergyLevelSensorService.getCharacteristic(Characteristic.ContactSensorState)
+                                    .onGet(async () => {
+                                        const state = this.consumptionNetEnergyLevelDetected;
+                                        const info = this.disableLogInfo ? false : this.emit('message', `Consumption net energy level sensor: ${state ? 'Active' : 'Not active'}`);
+                                        return state;
+                                    });
+                                accessory.addService(this.consumptionNetEnergyLevelSensorService);
+                            };
                         };
                     }
                 }
@@ -4057,6 +4262,34 @@ class EnvoyDevice extends EventEmitter {
                     this.ensembleStatusService.push(enphaseEnsembleStatusService);
                     accessory.addService(this.ensembleStatusService[0]);
                 }
+
+                //encharge grid state sensor service
+                if (this.enchargeGridStateSensor) {
+                    const debug = this.enableDebugMode ? this.emit('debug', `Prepare encharge grid state sensor service`) : false;
+                    this.enchargeGridStateSensorService = new Service.ContactSensor(`${accessoryName} Encharge Grid State`, `Encharge Grid State`);
+                    this.enchargeGridStateSensorService.getCharacteristic(Characteristic.ConfiguredName, `${accessoryName} Encharge Grid State`);
+                    this.enchargeGridStateSensorService.getCharacteristic(Characteristic.ContactSensorState)
+                        .onGet(async () => {
+                            const state = this.enchargerGridState = false;;
+                            const info = this.disableLogInfo ? false : this.emit('message', `Encharge grid state sensor: ${state ? 'Active' : 'Not active'}`);
+                            return state;
+                        });
+                    accessory.addService(this.enchargeGridStateSensorService);
+                };
+
+                //solar grid state sensor service
+                if (this.solarGridStateSensor) {
+                    const debug = this.enableDebugMode ? this.emit('debug', `Prepare solar grid state sensor service`) : false;
+                    this.solarGridStateSensorService = new Service.ContactSensor(`${accessoryName} Solar Grid State`, `Solar Grid State`);
+                    this.solarGridStateSensorService.getCharacteristic(Characteristic.ConfiguredName, `${accessoryName} Solar Grid State`);
+                    this.solarGridStateSensorService.getCharacteristic(Characteristic.ContactSensorState)
+                        .onGet(async () => {
+                            const state = this.enchargerGridState = false;;
+                            const info = this.disableLogInfo ? false : this.emit('message', `Solar grid state sensor: ${state ? 'Active' : 'Not active'}`);
+                            return state;
+                        });
+                    accessory.addService(this.solarGridStateSensorService);
+                };
 
                 //wireless connektion kit
                 if (wirelessConnectionKitInstalled) {